Add confirmation dialogs for dangerous actions.

+++ src/app/orders/orders/orders.component.ts
@@ -2,6 +2,8 @@ import { ToastrService } from 'ngx-toastr';
 import { Component, OnInit } from '@angular/core';
 import { Order } from '../order.model';
 import { OrdersService } from '../orders.service';
+import { MatDialog } from '@angular/material/dialog';
+import { ConfirmationDialogComponent } from 'src/app/confirmation-dialog/confirmation-dialog.component';
 
 @Component({
   selector: 'app-orders',
@@ -15,16 +17,26 @@ export class OrdersComponent implements OnInit {
 
   constructor(
     private orderService: OrdersService,
-    private toastService: ToastrService) { }
+    private toastService: ToastrService,
+    private dialog: MatDialog) { }
 
   ngOnInit(): void {
     this.orderService.getAll().subscribe(response => this.orders = response);
   }
 
   deleteById(id: number): void {
-    this.orderService.deleteById(id).subscribe(response => {
-      this.orders = response;
-      this.toastService.success("Item removed.");
+    const dialog = this.dialog.open(ConfirmationDialogComponent, {
+      data: {
+        content: "remove this order permanently"
+      }
+    });
+    dialog.afterClosed().subscribe(result => {
+      if (result) {
+        this.orderService.deleteById(id).subscribe(response => {
+          this.orders = response;
+          this.toastService.success("Item removed.");
+        });
+      }
     });
   }
