+++ src/app/shopping-cart/shopping-cart/shopping-cart.component.ts
@@ -3,6 +3,8 @@ import { ShoppingCart } from './../shopping-cart.model';
 import { Component, OnInit } from '@angular/core';
 import { ToastrService } from 'ngx-toastr';
 import { Router } from '@angular/router';
+import { MatDialog } from '@angular/material/dialog';
+import { ConfirmationDialogComponent } from 'src/app/confirmation-dialog/confirmation-dialog.component';
 
 @Component({
   selector: 'app-shopping-cart',
@@ -17,30 +19,49 @@ export class ShoppingCartComponent implements OnInit {
   constructor(
     private shoppingCartService: ShoppingCartService, 
     private toastService: ToastrService,
-    private router: Router) { }
+    private router: Router,
+    private dialog: MatDialog) { }
 
   ngOnInit(): void {
   }
 
   buy(): void {
-    this.shoppingCartService.buy().subscribe(
-      response => {
-        this.shoppingCartService.shoppingCart = response;
-        this.shoppingCart = this.shoppingCartService.shoppingCart;
-        this.toastService.success("Thank you for buying.");
-        this.router.navigate(['']);
+    const dialog = this.dialog.open(ConfirmationDialogComponent, {
+      data: {
+        content: "proceed with the payment"
       }
-    );
+    });
+    dialog.afterClosed().subscribe(result => {
+      if (result) {
+        this.shoppingCartService.buy().subscribe(
+          response => {
+            this.shoppingCartService.shoppingCart = response;
+            this.shoppingCart = this.shoppingCartService.shoppingCart;
+            this.toastService.success("Thank you for buying.");
+            this.router.navigate(['']);
+          }
+        ); 
+      }
+    });
   }
 
   removeProduct(productId: number): void {
-    this.shoppingCartService.removeProduct(productId).subscribe(
-      response => {
-        this.shoppingCartService.shoppingCart = response;
-        this.shoppingCart = this.shoppingCartService.shoppingCart;
-        this.toastService.success("Item removed.");
+    const dialog = this.dialog.open(ConfirmationDialogComponent, {
+      data: {
+        content: "remove this product from the shopping cart"
       }
-    );
+    });
+    dialog.afterClosed().subscribe(result => {
+      if (result) {
+        this.shoppingCartService.removeProduct(productId).subscribe(
+          response => {
+            this.shoppingCartService.shoppingCart = response;
+            this.shoppingCart = this.shoppingCartService.shoppingCart;
+            this.toastService.success("Item removed.");
+          }
+        );
+      }
+    });
   }
 
 }
